requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.36
pandas==2.2.2
numpy==2.0.1
scikit-learn==1.5.2
prophet==1.1.5        # for forecasting
joblib==1.4.2         # for model persistence
python-dotenv==1.0.1
